crypto/padmac/src/requirements.txt
pycrypto==2.6.1
Flask==2.1.1
uWSGI==2.0.20